requirements.txt
annotated-types==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.12" and python_version < "4.0"
basicauth==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.8.1 ; python_version >= "3.12" and python_version < "4"
certifi==2024.2.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4"
cloudevents==1.10.1 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4" and platform_system == "Windows"
deprecation==2.1.0 ; python_version >= "3.12" and python_version < "4"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
flask==3.0.3 ; python_version >= "3.12" and python_version < "4"
functions-framework==3.5.0 ; python_version >= "3.12" and python_version < "4"
gunicorn==22.0.0 ; python_version >= "3.12" and python_version < "4" and platform_system != "Windows"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4"
jinja2==3.1.3 ; python_version >= "3.12" and python_version < "4"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4"
neo4j-rust-ext==5.20.0.0 ; python_version >= "3.12" and python_version < "4.0"
openai==1.25.1 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.12" and python_version < "4"
pydantic-core==2.18.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.7.1 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.4 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.12" and python_version < "4.0"
watchdog==4.0.0 ; python_version >= "3.12" and python_version < "4"
werkzeug==3.0.2 ; python_version >= "3.12" and python_version < "4"
